refactor(Rule) : ScoreBoard Rule에서 관리
